+++ src/Ampersand/Session.php
@@ -17,6 +17,7 @@ use Ampersand\Misc\Config;
 use Psr\Log\LoggerInterface;
 use Ampersand\Core\Link;
 use Ampersand\Core\Relation;
+use Ampersand\Log\Logger;
 
 /**
  * Class of session objects
@@ -81,6 +82,7 @@ class Session
 
     public function reset()
     {
+        $this->logger->debug("Reset session {$this->id}");
         $this->sessionAtom->delete(); // Delete Ampersand representation of session
         session_regenerate_id(); // Create new php session identifier
         $this->setId();
@@ -100,29 +102,27 @@ class Session
             // TODO: can be removed when meat-grinder populates this meta-relation by itself
             if (!Config::get('loginEnabled')) {
                 foreach (Role::getAllRoles() as $role) {
-                    $this->sessionAtom->link(Concept::makeRoleAtom($role->label), 'sessionAllowedRoles[SESSION*Role]')->add();
+                    $roleAtom = Concept::makeRoleAtom($role->label);
+                    $this->sessionAtom->link($roleAtom, 'sessionAllowedRoles[SESSION*Role]')->add();
+                    // Activate all allowed roles by default
+                    $this->toggleActiveRole($roleAtom, true);
                 }
             }
-
-            // Activate all allowed roles by default
-            foreach ($this->getSessionAllowedRoles() as $atom) {
-                $this->toggleActiveRole($atom, true);
-            }
         } else {
-            $experationTimeStamp = time() - Config::get('sessionExpirationTime');
-            $lastAccessTime = $this->sessionAtom->getLinks('lastAccess[SESSION*DateTime]'); // lastAccess is UNI, therefore we expect max one DateTime from getLinks()
-            
-            // strtotime() returns Unix timestamp of lastAccessTime (in UTC). time() does also. Those can be compared
-            if (count($lastAccessTime) && strtotime(current($lastAccessTime)->tgt()->getLabel()) < $experationTimeStamp) {
+            if (isset($_SESSION['lastAccess']) && (time() - $_SESSION['lastAccess'] > Config::get('sessionExpirationTime'))) {
                 $this->logger->debug("Session expired");
-                // if(Config::get('loginEnabled')) \Ampersand\Log\Logger::getUserLogger()->warning("Your session has expired, please login again");
+                Logger::getUserLogger()->warning("Your session has expired");
                 $this->reset();
                 return;
             }
         }
         
-        // Set lastAccess time
-        $this->sessionAtom->link(date(DATE_ATOM), 'lastAccess[SESSION*DateTime]', false)->add();
+        // Update session variable. This is needed because windows platform doesn't seem to update the read time of the session file
+        // which will cause a php session timeout after the default timeout of (24min), regardless of user activity. By updating the
+        // session file (updating 'lastAccess' variable) we ensure the the session file timestamps are updated on every request. 
+        $_SESSION['lastAccess'] = time();
+        // Update lastAccess time also in plug/database to allow to use this aspect in Ampersand models
+        $this->sessionAtom->link(date(DATE_ATOM, $_SESSION['lastAccess']), 'lastAccess[SESSION*DateTime]', false)->add();
         
         Transaction::getCurrentTransaction()->runExecEngine()->close();
     }